+++ shop/util/cart.py
@@ -3,11 +3,14 @@ from shop.models.cartmodel import Cart
 from django.contrib.auth.models import AnonymousUser
 
 
-def get_or_create_cart(request):
+def get_or_create_cart(request, save=False):
     """
-    Let's inspect the request for session or for user, then either find a
-    matching cart and return it or create a new one bound to the user (if one
-    exists), or to the session.
+    Return cart for current visitor.
+
+    Let's inspect the request for user or session, then either find a
+    matching cart and return it or build a new one.
+
+    If ``save`` is True, cart object will be explicitely saved.
     """
     cart = None
     if not hasattr(request, '_cart'):
@@ -15,7 +18,7 @@ def get_or_create_cart(request):
             # There is a logged in user
             cart = Cart.objects.filter(user=request.user)  # a list
             if not cart:  # if list is empty
-                cart = Cart.objects.create(user=request.user)
+                cart = Cart(user=request.user)
             else:
                 cart = cart[0]  # Get the first one from the list
         else:
@@ -29,8 +32,10 @@ def get_or_create_cart(request):
                     except Cart.DoesNotExist:
                         cart = None
                 if not cart:
-                    cart = Cart.objects.create()
-                    session['cart_id'] = cart.id
+                    cart = Cart()
+        if save and not cart.pk:
+            cart.save()
+            request.session['cart_id'] = cart.id
         setattr(request, '_cart', cart)
     cart = getattr(request, '_cart')  # There we *must* have a cart
     return cart